requirements.txt
python-telegram-bot
opencv-python
pyautogui
numpy
python-dotenv